django/rest/drf_permissions/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
captcha==0.2.4
celery==3.1.25
celery-with-redis==3.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11
django-celery==3.2.1
django-cors-headers==2.3.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-haystack==2.6.1
django-redis==4.4.0
django-redis-cache==1.7.1
django-reversion==2.0.13
django-tinymce==2.7.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.3.1
future==0.16.0
gevent==1.3.4
greenlet==0.4.13
httplib2==0.11.3
idna==2.6
itsdangerous==0.24
itypes==1.1.0
jieba==0.38
Jinja2==2.10
jsonpath==0.75
kombu==3.0.37
lxml==4.2.1
Markdown==2.6.8
MarkupSafe==1.0
Pillow==5.1.0
pycparser==2.18
pycryptodome==3.6.3
Pygments==2.2.0
PyJWT==1.6.4
PyMySQL==0.8.1
pytz==2018.4
raven==6.9.0
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.10
tenacity==4.12.0
threadpool==1.3.2
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.14.1
Whoosh==2.7.4
XlsxWriter==1.0.5
xlwt==1.3.0
